Extract the work experience preview block into a separate WorkExperience.jsx file.
